requirements.txt
openai
tweepy
gradio
python-dotenv